Add functions to read and write color and gray PNG
There are functions to read a PNG from a file and get a matrix
struct back, and to write it back to memory. There are also
functions to convert between RGBA and greyscale images.

NB: the matrix type has changed, data is now a linear array
and not a two-dimensional one, because this is what lodepng
gives us back. It's a bit more arithmetic, but that's not
a big deal (and that makes less pointers).